refactor(fsd): features の hooks/ を model/ へ移設する (#240)
FSD §1-3「hook は model/ に置く」への是正（A1）。fleet の codemod
`nene2-a1-hooks-to-model`（nene2-standards 1.1.0）を適用。fleet 裁定 =
(a) ディレクトリ移動のみ（リネームなし）・vault が pilot。手作業移設なし。

- 移動: 10ファイル（5フック + 5 test 同居）を hooks/ → model/（純粋 rename・
  内容不変）。4 feature とも model/ 新規。
- import 書換: 7ファイル・8箇所（barrel 2 + ui→hooks 相対 6）を hooks/→model/
  の1トークンで codemod が追随。
- 対象外: shared/i18n/use-translation.ts（shared segment・A1 非対象）。

適用手順（証拠3点）:
- dry-run plan: move 10 / movedHooks 5（本文は PR body 参照）。dry-run の
  「rewrite 0」表示は集計しない仕様で偽り→実適用は rewrite 7。
- npm run check --prefix frontend **全体**緑（tsc / eslint --max-warnings 0 /
  format / vitest 221 passed / knip / build-storybook）。
- move-only diff（hook 本体は純粋 rename・importer は path token のみ）。
