+++ frontend/src/features/document-detail/index.ts
@@ -2,4 +2,4 @@ export { VoidModal } from './ui/VoidModal';
 export { RestoreModal } from './ui/RestoreModal';
 export { MetadataEditModal } from './ui/MetadataEditModal';
 export { DocumentHistoryTable } from './ui/DocumentHistoryTable';
-export type { OcrPrefill } from './hooks/use-metadata-edit';
+export type { OcrPrefill } from './model/use-metadata-edit';
