+++ frontend/src/features/document-detail/model/use-void-document.ts
