+++ frontend/src/features/document-search/index.ts
@@ -1,3 +1,3 @@
-export { useDocumentSearch } from './hooks/use-document-search';
+export { useDocumentSearch } from './model/use-document-search';
 export { DocumentSearchForm } from './ui/DocumentSearchForm';
 export { DocumentTable } from './ui/DocumentTable';
