+++ frontend/src/features/document-upload/ui/DocumentUploadModal.tsx
@@ -1,7 +1,7 @@
 import { env } from '@/shared/config/env';
 import { useTranslation } from '@/shared/i18n/use-translation';
 import { Button, Field, Input, Modal, Select } from '@/shared/ui';
-import { useDocumentUpload } from '../hooks/use-document-upload';
+import { useDocumentUpload } from '../model/use-document-upload';
 
 const CATEGORIES = ['invoice_received', 'contract', 'receipt', 'delivery_note', 'other'] as const;
